📦 refactor(plsql): Centralize code cleaning functionality
Move clean_code_and_map_literals function from extraction_workflow.py to dedicated
utils/code_cleaner.py module. Update CallDetailExtractor to consume pre-cleaned code
and literal mapping, removing redundant preprocessing logic. Update imports and
docstrings across affected files.

This change improves code organization by:
- Centralizing the string literal replacement and comment removal logic
- Avoiding redundant preprocessing in the call extraction workflow
- Adding better type annotations and documentation
- Removing duplicated code between modules
